# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Facebook Webhook + Kafka API",
    "version": "2.0.0",
    "description": "API xử lý thời gian thực với Facebook Webhook và Kafka"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "paths": {
    "/webhook": {
      "get": {
        "summary": "Xác minh Webhook với Facebook",
        "description": "Facebook gọi endpoint này để verify webhook bằng hub.verify_token",
        "parameters": [
          {
            "name": "hub.mode",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "subscribe"
          },
          {
            "name": "hub.verify_token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "abc123"
          },
          {
            "name": "hub.challenge",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "123456"
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook verified successfully"
          },
          "403": {
            "description": "Verify token không hợp lệ"
          }
        }
      },
      "post": {
        "summary": "Nhận sự kiện realtime từ Facebook Webhook",
        "description": "Nhận event comment/post từ Facebook, normalize payload và publish vào Kafka topic raw_events",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "object": "page",
                  "entry": [
                    {
                      "id": "123456789",
                      "time": 1714000000,
                      "changes": [
                        {
                          "field": "feed",
                          "value": {
                            "item": "comment",
                            "verb": "add",
                            "comment_id": "987654321",
                            "message": "Test comment from Swagger"
                          }
                        }
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook event received and pushed to Kafka successfully"
          }
        }
      }
    },
    "/api/page/{pageId}": {
      "get": {
        "summary": "Lấy thông tin Facebook Page",
        "parameters": [
          {
            "name": "pageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy thông tin Page thành công"
          }
        }
      }
    },
    "/api/page/{pageId}/posts": {
      "get": {
        "summary": "Lấy danh sách bài viết của Page",
        "parameters": [
          {
            "name": "pageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách bài viết thành công"
          }
        }
      },
      "post": {
        "summary": "Đăng bài viết mới lên Page",
        "parameters": [
          {
            "name": "pageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Bài viết test từ Swagger API"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đăng bài viết thành công"
          }
        }
      }
    },
    "/api/page/post/{postId}": {
      "delete": {
        "summary": "Xóa bài viết",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa bài viết thành công"
          }
        }
      }
    },
    "/api/page/post/{postId}/comments": {
      "get": {
        "summary": "Lấy danh sách comments của bài viết",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy comments thành công"
          }
        }
      }
    },
    "/api/page/post/{postId}/likes": {
      "get": {
        "summary": "Lấy danh sách likes của bài viết",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy likes thành công"
          }
        }
      }
    },
    "/api/page/{pageId}/insights": {
      "get": {
        "summary": "Lấy insights của Facebook Page",
        "parameters": [
          {
            "name": "pageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy insights thành công"
          }
        }
      }
    }
  }
}